Unlock PostgreSQL database
Fixes an issue where the single mutex in each database
using the PostgreSQL engine is locked while attempting to connect to
PostgreSQL and while running queries.

This is a problem because the same lock is held while reading the
database name and its comment field. This means that a single
failing service integration could make it almost impossible to read
the `system.databases` metadata table, which is used for monitoring
and other critical operations.

This would also make it really hard to drop the problematic integration
database, since the drop operation also does PostgreSQL queries.

The `cached` option of the PostgreSQL database engine does not
improve the situation :
It only caches the table structure but not the table existence, this is
why drop/attach/detach/list operations all need PostgreSQL queries.

This commit was applied from the patch file 0029-Unlock_PostgreSQL_Database.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>