Restrict "SHOW CREATE database"
Main service users are allowed the SHOW DATABASES access because it is
necessary for their operation. It is implicitly granted by ClickHouse when
giving access to a database.

However, we do not want to give them access to SHOW CREATE DATABASE.
This undocumented but enabled query shows the entire create statement,
unredacted. This is actually a useful feature for superusers, but can leak
credentials to other users.

Also only show the create query in system.tables to users that were able
to create that table.

This commit was applied from the patch files
0032-Restrict_Show_Create_Database.patch and 0033-Restrict_Create_Table_Query.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>