Fix ClickHouse restart with replicated table
ClickHouse does not inject the `{shard}` macro value when re-attaching
a table when the server starts.

It used to work in 21.x because we had the custom `{shard_name}`
macro which handled this case.

Now that this patch is removed, we have the `{shard}` bug: a table
can be created both directly and indirectly from the replicated
queries log in ZooKeeper, everything works until the server is
restarted.

When the server restarts, the shard information is not added to
the macro expansion context and the `ATTACH` fails with:

  Caught exception while loading metadata: Code: 62. DB::Exception:
  No macro 'shard' in config while processing substitutions in
  '/clickhouse/tables/{uuid}/{shard}'

Confusingly, `is_replicated_database` is only true for secondary
queries (those coming from the replicated queries log), that why
the extra condition needs to be added to handle the `ATTACH`.

This commit was applied from the patch file 0039-Attach_replicated_table.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>