Fix ClickHouse settings when adding a replica
ClickHouse global settings can be temporarily modified during a
single session or during a single query.

These settings can alter the behavior of a query, such as allowing
optional or experimental features when creating a table: this is used
for the JSON and Geo data types.

Although they affect table creation, these global settings are
different from the table settings.

When using a Distributed database, a replicated query must use
the same global settings on all servers: the query needs to have
the same effect on all servers.

To be able to do that, ClickHouse adds the global settings to each
entry in the DDL replication log. All servers can read the log and
execute the same query with the same settings.

However, when adding a replica to an existing cluster, the new
replica does not read the replication log. Instead, the replica reads
the table metadata in ZooKeeper. This is done in a function called
`recoverLostReplica` ("lost" means too out-of-sync to apply the
available DDL log entries, which is true for a new replica).

This table metadata contains the table settings but not the global
settings.

Due to that, the new replica will execute all the `CREATE TABLE`
queries with the table settings but without the global settings.

Because of this discrepancy when handling global settings, an
existing replica can create a table using a global experimental
flag, but a new replica will fail to create the same table.

Ideally, we'd apply all global settings when recovering a lost
replica, but that information is not available in the table metadata.

Instead, we can detect the queries executed by `recoverLostReplica`
and selectively disable the guardrails preventing the creation of
the table (these guardrails are pointless anyway since the table
already exists in all other replicas).

Since ClickHouse 23.8, most settings are correctly enabled, so this
patch became much smaller, but at least one setting is still missing.

Similarly, If the table was initially created with flatten_nested=0,
this caused a failure when adding a new replica to an existing
cluster: the table would be created with flatten_nested=1 on the new
replica and the schemas would not match.

This commit was applied from the patch file 0040-Recover_replica_settings.patch