Fix unbounded replication queue
If a replica is struggling to execute all tasks in its replication
queue, the queue can grow to a huge size and is even slower to manage
when its larger (it's an `std::list`).

This queue is also present in ZooKeeper, for each replica.

There is also a shared queue, which is called the log. It's used to
communicate changes between replicas. One replica will push an item on
the log as part of executing a query, then all other replicas will read
that log item and repush it in their own queue.

(This is an approximation, the queue items are not identical to the log
items, but filling the log fills the queues).

Each replica monitors the log size and which replicas are deemed "lost"
when they are too late to consume log items. However, the replicas do
not monitor the size of the queue of the other replicas.

If a replica is pushing lots of items in the log and another replica is
able to copy them to their own queue, but too slow to consume the queue
items, then we have a problem.

At some point, if the imbalance persists, ZooKeeper will crash because
it won't have enough memory to hold all the queues. It will be very hard
to recover: ZooKeeper itself often corrupt its state when OOMing.
Manually cleaning hundred of thousand of items in the queues to know
which ones can be removed and which one are necessary to keep the
replicated data in a consistent state is almost impossible.

To fix that, we reuse the infrastructure that delays insert queries when
there are too many parts. The same logic is applied when any replica queue
size is too large for the table, or the grand total of all replicas' queue
sizes, over all tables, is too large. We also do the same to fail queries
if either counters reach a second threshold, exactly like it's done with
the parts count.

In each replica, a thread is updating a local copy of the maximum queue
size in all replicas. A map is added to the context to keep track of these
per-storage maxima. These values are then used to gate the inserts. We do
it like that to avoid adding many ZooKeeper queries in the hot path, and we
don't need a very accurate queue size to get the feedback loop we need.

This commit was applied from the patch file 0050-Fix_unbounded_replication_queue.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>