Check MergeTree settings constraints before enqueuing DDL queries
The constraints were previously only checked when executing the query
from the replicated DDL queue. At that point the DDLWorker was using
the system profile for settings constraints.

This meant that a non-admin user, using the default profile, could use
a MergeTree settings that was not allowed by its profile.

Fix that by checking the MergeTree settings constraints before the
query is enqueued, when we still have the query context attached to
the user running the query.

We check both Replicated*MergeTree and apparently non-replicated
MergeTree because this check happens before we rewrite the query
to enforce Replicated*MergeTree engine types.

This commit was applied from the patch file 0057-Early_MergeTree_settings_constraints.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>