Add Backup disk type
This adds a new disk type that wraps an object storage disk and
enables coordination with our backup system.

When this disk type is used as a wrapper on top of an object storage
disk, the files in object storage are not really deleted.

Instead, a local file is created for each file in object storage that
should be deleted.

Our backup system will then be in charge to do garbage collection
and only delete files are are not referenced by any backup.

The file names of the deletion markers are like this:
CLICKHOUSE_DIR/disks/remote_backup/backup/ESCAPED_OBJECT_PATH

The object path is escaped to simplify directory management and uses
the same escape mechanism that ClickHouse uses in many places to
reference files (essentially escape encoding).

In case Backup disk is wrapping a Cache disk then it lets the remove
calls pass

This commit was applied from the patch file 0059-Backup_disk.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>