Fix ClickHouse trying to read non-existent certificate files
If both the certificate file and key file are defined, even when they are
not used, ClickHouse will try to read and parse them. We can't even
work around the issue by creating empty files.

There were multiple iterations of improvements for this patch.

The latest approach to the problem is to make sure there's one
successful call for `tryLoad` before calling `tryReloadAll`. During the
research it was noticed that `tryLoad` loads exactly one prefix
`"openSSL.server."` and there's no place in the code that adds another
prefix (at least in our usecase). This means that `tryReloadAll` acts
effectively the same as `tryLoad`.

This commit was applied from the patch file 0064-Lazy_certificates.patch

[DDB-1898]

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>
Co-authored-by: Joe Lynch <joe.lynch@aiven.io>
Co-authored-by: Aris Tritas <aris.tritas@aiven.io>
Co-authored-by: Tilman Möller <tilman.moeller@aiven.io>