Allow delegating S3 signature to a separate process
The process is accessed over HTTP and the URL to access it can be
configured with the signature_delegation_url parameter in the S3 disk
configuration.

ClickHouse will make a POST request with a json body:
{
    "canonicalRequest": "PUT\n...",
}
And expects a JSON response with status 200 and this structure:
{
    "signature": 01234567890abcdefg..."
}

The canonical request matches the format defined by AWS for
signatures and all the extra parameters allow computing the full
string to sign.

The canonical request also contains all the required  information
(path, host, operation...) to let the proxy decide of the request is
allowed and can be signed. It's also enough information to know
the class and cost of the request.

This commit was applied from the patch file 0066-Delegated_signature.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>