Add per-server override for max bytes to merge/mutate
During a maintenance upgrade, we do not wait for the completion of
merges and mutations since they are not required to make sure we have
all the data from the previous nodes.

We do wait for `GET_PART` and similar tasks, since we need the new
nodes to get parts from the old nodes. This is implemented using
`SYSTEM SYNC REPLICA ... LIGHTWEIGHT`.

Some merges and mutations prevent the execution of `GET_PART` tasks
that overlap the range of the merge or the mutation. So, even if we
are not waiting for `MERGE_PARTS` tasks, these tasks can slow down
the completion of a maintenance upgrade.

We need to execute `MERGE_PARTS` tasks to avoid having too many
parts in the same partition or table. We also need to execute the TTL
delete rules and ensure disk usage does not grow too much (TTL
deletes are implemented as a subtype of the merge tasks).

A possible tradeoff is to only execute merge tasks if they are not too large,
the small merges are the ones that keep the number of parts low when
the are many small INSERT queries.

We can't use the normal merge tree settings like
`max_bytes_to_merge_at_min/max_space_in_pool` because they would be
applied with `ALTER TABLE` which would step onto user-managed objects
and become persisted as part of the table definition. They are also
replicated, we can't set them only on new nodes.

The patch implements per-server overrides, that we can use to limit
the size of merge and mutate tasks, both when a node creates new
tasks and when it decides which task to execute.

This is only usable with a `LIGHTWEIGHT` sync: if we don't execute
some tasks, we also need to not wait for them, or we would wait
forever. `SYNC REPLICA` with the `LIGHTWEIGHT` flag allows us to
not wait for these tasks.

This commit was applied from the patch file 0078-Global_merge_and_mutate_override.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>