Properly fix 0064-Lazy_certificates.patch
Previous fix was incorrect as statement
`!config.has("tcp_port_secure") && !config.has("https_port")` was always
true, even when certificates were not ready yet. This reintroduced the
problem when ClickHouse tried to load certificates and errored out with
"No such file or directory", which the original patch tried to fix.

The proper approach to the problem is to make sure there's one
successful call for `tryLoad` before calling `tryReloadAll`. During the
research it was noticed that `tryLoad` loads exactly one prefix
`"openSSL.server."` and there's no place in the code that adds another
prefix (at least in our usecase). This means that `tryReloadAll` acts
effectively the same as `tryLoad`.

This commit restores the original if-statement thus fixing the error
while loading the non-existent certificates. Additionally, it replaces
the call for `tryReloadAll` from `main_config_reloader` with the call
for `tryLoad` for the reasons explained above.

This commit has been tested in the test environment and proved to be
working.

[DDB-1898]

Co-authored-by: Aris Tritas <aris.tritas@aiven.io>
Co-authored-by: Tilman Möller <tilman.moeller@aiven.io>