Allow refreshable mat views when using zk (rather than ch keeper)
There is only one multiRead which happens when fetching the paths
`keeper_path`, `running` and `paused`. If we allow a non multi read
here, then we fetch each path one at a time, which could be not
consistent. But this is no issue for `running` since we try to create an
ephemeral node on this path later anyway (which will allow only one
replica to create the path and run the refresh). `paused` is also okay,
it might just mean that we miss an early refresh if it is unpaused, or we
accidentally refresh a final time after a pause.