Restrict SHOW CREATE DATABASE access
Main service users are allowed the SHOW DATABASES access because it is necessary for their operation. It is implicitly granted by ClickHouse when giving access to a database.

However, we do not want to give them access to SHOW CREATE DATABASE.
This query shows the entire create statement, unredacted. This is actually a useful feature for superusers, but can leak credentials to other users.

Also only show the create query in system.tables to users that were able to create that table.

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>