Added support for protected users
Add an extra flag to each user, called "Protected".

Users with the "Protected" flag can only be created, altered or removed
by users with an extra `PROTECTED_ACCESS_MANAGEMENT` privilege.

This privilege is only meant to be granted to our internal admin user,
and the "Protected" flag will be given to the main service user,
allowing it to exist as an SQL user (which is necessary to have less privileges than
an hardcoded XML users), but not be removable or alterable by the
cluster users.

Prevent users from dropping themselves, even if they have PROTECTED_ACCESS_MANAGEMENT

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>