Add Backup disk type
This adds a new disk type that wraps an object storage disk and
enables coordination with our backup system. When this disk type
is used as a wrapper on top of an object storage disk, the files
in object storage are not really deleted. Instead, a local file is
created for each file in object storage that should be deleted.

Our backup system will then be in charge to do garbage collection
and only delete files that are not referenced by any backup.

The file names of the deletion markers are like this:
CLICKHOUSE_DIR/disks/remote_backup/backup/ESCAPED_OBJECT_PATH

The object path is escaped to simplify directory management and uses
the same escape mechanism that ClickHouse uses in many places to
reference files (essentially escape encoding).

In case Backup disk is wrapping a Cache disk then it lets the remove
calls pass.

Changes:
- Added BackupObjectStorage wrapper class that intercepts delete
  operations and creates local deletion marker files
- Added registerDiskBackup factory function to register the backup
  disk type
- Extended IObjectStorage interface with getLayerName() and
  getWrappedObjectStorage() methods for generic layer traversal
- Renamed getCacheLayersNames() to getLayersNames() to support both
  cache and backup layers
- Added supportsLayers() method to IDisk interface
- Updated DiskObjectStorage to support wrapping with backup layer
- Updated MergeTreeData to track backup layers in addition to cache
  layers

Adaptations from original patch:
- Removed cloneObjectStorage() method (not in current IObjectStorage
  interface)
- Fixed include order to prevent Poco namespace resolution issues
- Updated all wrapper classes (ReadOnlyDiskWrapper, DiskEncrypted) to
  support the new layer interface

This commit was applied from the patch file 0059-Backup_disk.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>