Fix uncaught exception if S3 storage fails
If an exception happened while copying data, then the finalize call was not done explicitly. The finalize was then be done while the out object is destructed and that led to a fatal error, killing the entire server.
To fix that, catch the exception that might happen while copyData is running, run finalize there and rethrow.
The rethrow is very important, the finalize after a failing copyData can also not-throw, in which case the initial error would be forgotten and ClickHouse would believe the copy succeeded.

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>